requirements.txt
discord.py
python-dotenv
yt-dlp
PyNaCl==1.4.0
